fixed ChatContent component errors
